Split Entity to partial classes
